Renamed four_momenta to four_momentum, since its singular.
