+++ ostrawling/event/__init__.py
@@ -1,3 +1,3 @@
 from .event import *
 from .particle import *
-from .four_momenta import *
+from .four_momentum import *
